Chapter 7 is now chapter 8
