+++ Chapter_8_Coding_JavaScript/Excercise8.01/Excercise8.01.js
@@ -1,10 +1,10 @@
 /*
-* Michael Harrison.
-* 08/12/2019
-*
-* Packt The MongoDB Workshop
-* Chapter 7
-*/
+ * Michael Harrison.
+ * 08/12/2019
+ *
+ * Packt The MongoDB Workshop
+ * Chapter 8
+ */
 
 // Excercise code for Packt MongoDB For Begginers.
 // This code can be run using NodeJS, make sure you have the MongoDB node driver installed with NPM before running.
@@ -16,16 +16,15 @@
 const MongoClient = require('mongodb').MongoClient;
 
 // Create a new url variable.
-const url = 'mongodb+srv://user:password@server-abcde.gcp.mongodb.net/test?retryWrites=true&w=majority';
+const url =
+  'mongodb+srv://user:password@server-abcde.gcp.mongodb.net/test?retryWrites=true&w=majority';
 
 // Create a new MongoClient.
 const client = new MongoClient(url);
 
 // Open the connection using the .connect function.
-client.connect(function(err) {
-    // Within the connection block, add a console.log to confirm the connection
-    console.log('Connected to MongoDB with NodeJS!');
-    client.close(); // Close our connection at the end.
-})
-
-
+client.connect(function (err) {
+  // Within the connection block, add a console.log to confirm the connection
+  console.log('Connected to MongoDB with NodeJS!');
+  client.close(); // Close our connection at the end.
+});
