+++ Chapter_8_Coding_JavaScript/Excercise8.03/Excercise8.03.js
@@ -3,39 +3,21 @@
  * 08/12/2019
  *
  * Packt The MongoDB Workshop
- * Chapter 7
+ * Chapter 8
  */
 
 // Excercise code for Packt MongoDB For Begginers.
 // This code can be run using NodeJS, make sure you have the MongoDB node driver installed with NPM before running.
 
-//// TOPIC D: "Advanced Queries with the Driver."
-// Excercise 4: Advanced Queries with the NodeJS Driver.
+//// TOPIC C: "Error handling with the Driver."
+// Excercise 3: Error Handling with the NodeJS Driver.
 const MongoClient = require('mongodb').MongoClient;
 const url =
   'mongodb+srv://user:password@server-abcde.gcp.mongodb.net/test?retryWrites=true&w=majority';
 const client = new MongoClient(url);
 const databaseName = 'sample_mflix';
-const collectionName = 'chapter7_excercise4';
-
-const updateName = function (client, database, oldName, newName) {
-  database
-    .collection(collectionName)
-    .updateOne({ name: oldName }, { $set: { name: newName } }, function (
-      err,
-      result
-    ) {
-      if (err) {
-        console.log('Error updating');
-        console.log(err);
-        client.close();
-        return false;
-      }
-      console.log('Updated documents #:');
-      console.log(result.modifiedCount);
-      client.close();
-    });
-};
+const collectionName = 'movies';
+const query = { genres: { $thisIsNotAnOperator: ['Romance'] } };
 
 // Open the connection using the .connect function.
 client.connect(function (err) {
@@ -53,8 +35,24 @@ client.connect(function (err) {
     console.log('Database object doesnt exist!');
     return false;
   }
-
-  updateName(client, database, 'Ned Stark', 'Greg Stark');
-  updateName(client, database, 'Robb Stark', 'Bob Stark');
-  updateName(client, database, 'Bran Stark', 'Brad Stark');
+  database
+    .collection(collectionName)
+    .find(query)
+    .limit(2)
+    .project({ title: 1 })
+    .each(function (err, doc) {
+      if (err) {
+        console.log('Query error.');
+        console.log(err);
+        client.close();
+        return false;
+      }
+      if (doc) {
+        console.log('Current doc');
+        console.log(doc);
+      } else {
+        client.close(); // Close our connection.
+        return false; // End the each loop.
+      }
+    });
 });
