+++ Chapter_8_Coding_JavaScript/Excercise8.05/Excercise8.05.js
@@ -3,7 +3,7 @@
 * 08/12/2019
 *
 * Packt The MongoDB Workshop
-* Chapter 7
+* Chapter 8
 */
 
 // Excercise code for Packt MongoDB For Begginers.
